feat: product previews
