+++ apps/sap-commerce-cloud/app-actions/src/constants.ts
@@ -1,1 +1,22 @@
 export const AIR_HEADER = 'xj823lbq';
+
+export const DEFAULT_FIELDS = [
+  'code',
+  'name',
+  'summary',
+  'price(formattedValue,DEFAULT)',
+  'images(galleryIndex,FULL)',
+  'averageRating',
+  'stock(DEFAULT)',
+  'description',
+  'availableForPickup',
+  'url',
+  'numberOfReviews',
+  'manufacturer',
+  'categories(FULL)',
+  'priceRange',
+  'multidimensional',
+  'configuratorType',
+  'configurable',
+  'tags',
+];
