+++ apps/sap-commerce-cloud/frontend/src/api/dataTransformers.ts
@@ -36,3 +36,22 @@ export const baseSiteTransformer =
   (item: Hash): string => {
     return get(item, ['uid'], '');
   };
+
+export const productDetailsTransformer =
+  ({ apiEndpoint }: ConfigurationParameters) =>
+  (item: Hash): Product => {
+    const id = get(item, ['id'], '');
+    let imageUrl = get(item, ['images', 0, 'url'], '');
+    if (imageUrl.length > 0) {
+      imageUrl = apiEndpoint + imageUrl;
+    }
+    return {
+      id,
+      image: imageUrl,
+      name: get(item, ['name'], '')
+        .replaceAll('<em class="search-results-highlight">', '')
+        .replaceAll('</em>', ''),
+      sku: get(item, ['code'], ''),
+      productUrl: '',
+    };
+  };
